+++ packages/common/src/entities/environments/authenticators/authenticator.types.ts
@@ -1,4 +1,4 @@
-import {IEnvironmentAuth} from '../../models';
+import {IEnvironmentAuth} from '../../../models';
 import {AUTHENTICATOR_TYPES} from './authenticators.constants';
 
 export interface IAuthenticator {
