+++ packages/common/src/entities/environments/authenticators/authenticators.constants.ts
