+++ packages/common/src/entities/environments/authenticators/index.ts
