+++ packages/common/src/entities/environments/environment.constants.ts
