+++ packages/common/src/entities/environments/oauth-utils.ts
