+++ packages/common/src/utils/dbmss/generate-plugin-certs.ts
@@ -4,7 +4,7 @@ import {
     NEO4J_JWT_CERT_ATTRS,
     NEO4J_JWT_CERT_BIT_LENGTH,
     NEO4J_JWT_CERT_VALIDITY_YEARS,
-} from '../../environments/environment.constants';
+} from '../../entities/environments/environment.constants';
 
 export function generatePluginCerts(passphrase: string): {publicKey: string; privateKey: string} {
     // generate a keypair or use one you have already
