+++ packages/common/src/utils/dbmss/index.ts
@@ -6,3 +6,4 @@ export * from './dbms-versions';
 export * from './generate-plugin-certs';
 export * from './download-neo4j';
 export * from './extract-neo4j';
+export {parseNeo4jConfigPort} from './parse-neo4j-config-port';
