+++ packages/common/src/utils/dbmss/neo4j-admin-cmd.ts
@@ -2,7 +2,7 @@ import fse from 'fs-extra';
 import path from 'path';
 
 import {NotFoundError, DependencyError} from '../../errors';
-import {NEO4J_BIN_DIR, NEO4J_ADMIN_BIN_FILE} from '../../environments/environment.constants';
+import {NEO4J_BIN_DIR, NEO4J_ADMIN_BIN_FILE} from '../../entities/environments/environment.constants';
 import {resolveRelateJavaHome} from './resolve-java';
 import {spawnPromise} from './spawn-promise';
 import {EnvVars} from '../env-vars';
