+++ packages/common/src/utils/index.ts
@@ -1,4 +1,3 @@
-export {parseNeo4jConfigPort} from './parse-neo4j-config-port';
 export {registerSystemAccessToken, getSystemAccessToken} from './system-access-token';
 export {envPaths} from './env-paths';
 export {download} from './download';
