+++ packages/common/src/utils/system/create-environment-instance.ts
@@ -1,7 +1,7 @@
 import {EnvironmentConfigModel} from '../../models';
 import {InvalidConfigError} from '../../errors';
 
-import {EnvironmentAbstract, RemoteEnvironment, LocalEnvironment, ENVIRONMENT_TYPES} from '../../environments';
+import {EnvironmentAbstract, RemoteEnvironment, LocalEnvironment, ENVIRONMENT_TYPES} from '../../entities/environments';
 
 export async function createEnvironmentInstance(
     config: EnvironmentConfigModel,
